increased version to 0.501 Added Bijection, unicode package, Constants, squareInPlace for all numeric types.
refactored Gamma, Constant, and Unicode utils into their own files.

+++ vector/shared/src/main/scala/ai/dragonfly/math/Factorial.scala
@@ -132,7 +132,7 @@ object Factorial {
    * @return x!
    */
 
-  def apply(x:Int):BigInt = if (x < table.length) table(x.toInt) else partialFactorial(BigInt(x))
+  def apply(x:Int):BigInt = if (x < table.length) table(x) else partialFactorial(BigInt(x))
 
   /**
    * Computes x!, the factorial(x), of Long x.
